Remove role="tab" from mobile tabs and progressively enhance to button
Fixes https://github.com/alphagov/govuk-design-system/issues/1653

When running the Lighthouse accessibility audit, the tabs for HTML and Nunjucks on mobile were flagged because they
used `role="tab"` without being nested within a parent element with `role="tablist"`.

The 'tabs' on mobile actually behave more like an accordion and it's not possible to nest them within `ul` and `li` elements,
so it made more sense to remove `role="tab"`.

Because the 'tabs' on mobile don't actually link anywhere when JavaScript is enabled, we should progressively enhance
the links to become buttons. When JavaScript is disabled, the tabs remain as links and link directly to each code example.