Add versioning to consent cookie
Why
---

If we are likely to move to a sub-domain consent approach in the future,
we will need to 'reset' the cookie banner and consent cookie for users
who have previously consented as what they have consented to has now
changed. This also applies if in the future we add a new cookie category
or important new cookie.